# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PETUK Restaurant Business Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # ── MENU ──────────────────────────────────────────────────────────────────
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── ORDERS ────────────────────────────────────────────────────────────────
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Register new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/lookup:
    get:
      operationId: lookupMember
      tags: [members]
      summary: Lookup member by phone
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberLookup"
  /members/{id}:
    put:
      operationId: updateMember
      tags: [members]
      summary: Update member (admin — adjust points etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # ── COUPONS ───────────────────────────────────────────────────────────────
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Toggle coupon active/inactive (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── RIDERS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [id, username, displayName, role, station]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        station:
          type: string
    AdminUser:
      type: object
      required: [id, username, displayName, role, station, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        station:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
...
    MenuItem:
      type: object
      required: [id, name, price, category, isAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        category:
          type: string
        emoji:
          type: ["string", "null"]
        imagePath:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    MenuItemInput:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        emoji:
          type: string
        imagePath:
          type: string
        isAvailable:
          type: boolean

    # ── ORDERS ────────────────────────────────────────────────────────────
...
    Member:
      type: object
      required: [id, phone, name, points, tier, joined]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        points:
          type: integer
        tier:
          type: string
        discountPercent:
          type: number
        tierColor:
          type: string
        joined:
          type: string
    MemberLookup:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        member:
          $ref: "#/components/schemas/Member"
    MemberInput:
      type: object
      required: [phone, name]
      properties:
        phone:
          type: string
        name:
          type: string
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        points:
          type: integer

    # ── COUPONS ───────────────────────────────────────────────────────────
    Coupon:
      type: object
      required: [id, code, type, value, minOrder, maxUses, usedCount, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [fixed, percent]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        usedCount:
          type: integer
        expiry:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    CouponInput:
      type: object
      required: [code, type, value]
      properties:
        code:
          type: string
        type:
          type: string
          enum: [fixed, percent]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        expiry:
          type: string
    CouponUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        code:
          type: string
        type:
          type: string
          enum: [fixed, percent]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        expiry:
          type: string
    CouponValidateInput:
      type: object
      required: [code, subtotal]
      properties:
        code:
          type: string
        subtotal:
          type: number
    CouponValidation:
      type: object
      required: [valid, discount, message]
      properties:
        valid:
          type: boolean
        discount:
          type: number
        message:
          type: string

    # ── RIDERS ────────────────────────────────────────────────────────────
...