# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PETUK Restaurant Business Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  name?: string;
  points?: number;
}